backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyJWT==2.8.0
pymysql==1.1.0
bcrypt==4.1.2
python-dotenv==1.0.0
gunicorn==21.2.0
cryptography==41.0.7
